+++ src/QueryOperators.jl
@@ -3,10 +3,12 @@ module QueryOperators
 using DataStructures
 using IteratorInterfaceExtensions
 using TableShowUtils
+import DataValues
 
 export Grouping, key
 
 include("operators.jl")
+include("NamedTupleUtilities.jl")
 
 include("enumerable/enumerable.jl")
 include("enumerable/enumerable_groupby.jl")
@@ -21,6 +23,7 @@ include("enumerable/enumerable_defaultifempty.jl")
 include("enumerable/enumerable_count.jl")
 include("enumerable/enumerable_take.jl")
 include("enumerable/enumerable_drop.jl")
+include("enumerable/enumerable_unique.jl")
 include("enumerable/show.jl")
 
 include("source_iterable.jl")